♻️ [refactor/withdraw] CKPS-16 users 전체 폴더 구조 개선 리펙토링
